test: consolidate focused coverage
