+++ Tests/ArknightsClientTests/Shared/Support/GameConfigurationTests.swift
@@ -4,20 +4,16 @@ import Testing
 
 @testable import ArknightsClient
 
-@Test
-func byteSizeParsesWholeGigabytes() {
-	#expect(GameConfiguration.parseByteSize("30GB") == 30_000_000_000)
-}
-
-@Test
-func byteSizeParsesSpacedUnitAndDecimalValue() {
-	#expect(GameConfiguration.parseByteSize("38.5 GB") == 38_500_000_000)
-}
-
-@Test
-func byteSizeParsesOtherUnits() {
-	#expect(GameConfiguration.parseByteSize("512MB") == 512_000_000)
-	#expect(GameConfiguration.parseByteSize("2TB") == 2_000_000_000_000)
+@Test(
+	arguments: [
+		("30GB", Int64(30_000_000_000)),
+		("38.5 GB", Int64(38_500_000_000)),
+		("512MB", Int64(512_000_000)),
+		("2TB", Int64(2_000_000_000_000)),
+	]
+)
+func byteSizeParsesRecognizedUnits(input: String, expected: Int64) {
+	#expect(GameConfiguration.parseByteSize(input) == expected)
 }
 
 @Test
